[#1405] fixed warnings

+++ src/lib/dhcpsrv/tests/reservation_modes_parser_unittest.cc
@@ -62,7 +62,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": false, \n"
         "   \"reservations-out-of-pool\": false \n"
         "} \n",
-        Network::HR_DISABLED
+        static_cast<uint8_t>(Network::HR_DISABLED)
         },
         {
         "reservation modes global enabled",
@@ -71,7 +71,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": false, \n"
         "   \"reservations-out-of-pool\": false \n"
         "} \n",
-        Network::HR_GLOBAL
+        static_cast<uint8_t>(Network::HR_GLOBAL)
         },
         {
         "reservation modes in-subnet enabled",
@@ -80,7 +80,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": true, \n"
         "   \"reservations-out-of-pool\": false \n"
         "} \n",
-        Network::HR_IN_SUBNET
+        static_cast<uint8_t>(Network::HR_IN_SUBNET)
         },
         {
         "reservation modes global and in-subnet enabled",
@@ -89,7 +89,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": true, \n"
         "   \"reservations-out-of-pool\": false \n"
         "} \n",
-        Network::HR_GLOBAL|Network::HR_IN_SUBNET
+        static_cast<uint8_t>(Network::HR_GLOBAL|Network::HR_IN_SUBNET)
         },
         {
         "reservation modes out-of-pool enabled",
@@ -98,7 +98,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": false, \n"
         "   \"reservations-out-of-pool\": true \n"
         "} \n",
-        Network::HR_OUT_OF_POOL
+        static_cast<uint8_t>(Network::HR_OUT_OF_POOL)
         },
         {
         "reservation modes global and out-of-pool enabled",
@@ -107,7 +107,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": false, \n"
         "   \"reservations-out-of-pool\": true \n"
         "} \n",
-        Network::HR_GLOBAL|Network::HR_OUT_OF_POOL
+        static_cast<uint8_t>(Network::HR_GLOBAL|Network::HR_OUT_OF_POOL)
         },
         {
         "reservation modes in-subnet and out-of-pool enabled",
@@ -116,7 +116,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": true, \n"
         "   \"reservations-out-of-pool\": true \n"
         "} \n",
-        Network::HR_IN_SUBNET|Network::HR_OUT_OF_POOL
+        static_cast<uint8_t>(Network::HR_IN_SUBNET|Network::HR_OUT_OF_POOL)
         },
         {
         "reservation modes global, in-subnet and out-of-pool enabled",
@@ -125,28 +125,28 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": true, \n"
         "   \"reservations-out-of-pool\": true \n"
         "} \n",
-        Network::HR_GLOBAL|Network::HR_IN_SUBNET|Network::HR_OUT_OF_POOL
+        static_cast<uint8_t>(Network::HR_GLOBAL|Network::HR_IN_SUBNET|Network::HR_OUT_OF_POOL)
         },
         {
         "only global",
         "{ \n"
         "   \"reservations-global\": true \n"
         "} \n",
-        Network::HR_GLOBAL
+        static_cast<uint8_t>(Network::HR_GLOBAL)
         },
         {
         "only in-subnet",
         "{ \n"
         "   \"reservations-in-subnet\": true \n"
         "} \n",
-        Network::HR_IN_SUBNET
+        static_cast<uint8_t>(Network::HR_IN_SUBNET)
         },
         {
         "only out-of-pool",
         "{ \n"
         "   \"reservations-out-of-pool\": true \n"
         "} \n",
-        Network::HR_OUT_OF_POOL
+        static_cast<uint8_t>(Network::HR_OUT_OF_POOL)
         },
         {
         "only global and in-subnet",
@@ -154,7 +154,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-global\": true, \n"
         "   \"reservations-in-subnet\": true \n"
         "} \n",
-        Network::HR_GLOBAL|Network::HR_IN_SUBNET
+        static_cast<uint8_t>(Network::HR_GLOBAL|Network::HR_IN_SUBNET)
         },
         {
         "only global and out-of-pool",
@@ -162,7 +162,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-global\": true, \n"
         "   \"reservations-out-of-pool\": true \n"
         "} \n",
-        Network::HR_GLOBAL|Network::HR_OUT_OF_POOL
+        static_cast<uint8_t>(Network::HR_GLOBAL|Network::HR_OUT_OF_POOL)
         },
         {
         "only in-subnet and out-of-pool",
@@ -170,7 +170,7 @@ TEST_F(HostReservationModesParserTest, validContent) {
         "   \"reservations-in-subnet\": true, \n"
         "   \"reservations-out-of-pool\": true \n"
         "} \n",
-        Network::HR_IN_SUBNET|Network::HR_OUT_OF_POOL
+        static_cast<uint8_t>(Network::HR_IN_SUBNET|Network::HR_OUT_OF_POOL)
         },
     };
